move ble::Bounded in its own file.
